+++ src/ov_SC04_010/ov_SC04_010_jr_801380E0.c
@@ -2287,7 +2287,7 @@ DEFINE_func_8013EA54()  /* dedup: shared engine-core @0x8013EA54 (src/shared) */
 
 extern s32 func_80141C50(void);
 extern void func_8013F350(void);
-extern void func_8013FAF8(s32 a0, s32 a1);
+extern void func_8013FAF8(s16 a0, s16 a1);
 extern s16 func_8014168C(s16 a0);
 extern s32 func_8014032C(s32 a0, s32 a1);
 extern void func_80139954(void);
@@ -2366,7 +2366,7 @@ extern s16 D_8017F514;
 extern s16 D_8017F516;
 extern unsigned char D_8017F518;
 extern void func_801376E8(int a0, int a1);
-extern void func_8013FAF8(s32 a0, s32 a1);
+extern void func_8013FAF8(s16 a0, s16 a1);
 extern unsigned char *func_80141CA4(void);
 
 // @class: struct
